01_vulnerable_dependencies/requirements.txt
PyYAML==3.13 